requirements.txt
# Requirements for SecureChat
pycryptodome